Mérnöki átvizsgálás: hibajavítások, takarékosabb működés, minőségi kapuk
Hibák (mind javítva, tesztekkel):
- a célmappa hiányát a torrent hozzáadása UTÁN néztük indulásnál, így a
  libtorrent a leválasztott meghajtó helyén kezdhetett volna írni
- az időzített szünet lejártakor a folytatás kivételt dobhatott, ami a démont
  állította le; most hibaállapotba megy és fut tovább
- hibás torrent hozzáadása "szellem" munkát hagyott hátra, ami minden további
  letöltést blokkolt
- a kérésre indított ellenőrzés hiányzó célmappánál is elindult volna
- a sikeres ellenőrzés nem nullázta az ellenőrzési kísérletek számlálóját
- a "tiszta leállás" jelzés kivétel miatti kilépéskor is kikerült
- a szüneteltetés érvénytelen fogantyúnál elszállhatott
- a session-statisztika mezőit indexszel olvastam, a Python-kötés viszont
  névvel indexelt szótárat ad (a hozzá írt teszt bukott el rá)

Modernizálás és takarékosság:
- az elavult session.status() helyett post_session_stats() + session_stats_alert
- lemezoldali hangolás a libtorrent ajánlásai szerint: külön hash-szálak
  (gyorsabb ellenőrzés), nagyobb fájlgyorsítótár, kisebb peer-lista
- az ismeretlen beállításokat kiszűrjük, hogy más libtorrent-verzióval is induljon
- az állapotfájl csak valódi változáskor íródik, a haladás-mentés fsync nélkül
- a last.json a memóriában, nem minden lekérdezésnél olvassuk újra
- tétlenül ritkább ébredés; a .torrent átvitele base64-gyel a hex helyett
- a napló olvasása közös modulba került, gyorsítótárral (a felület és a
  parancssor is ezt használja); a parancssor eddig az egész naplót beolvasta
- kevesebb naplózaj (állapotváltás debug szinten, hibás darabok ritkítva)

Minőség:
- ruff (hibaminták, stílus, teljesítmény, biztonság) és mypy: mindkettő tiszta,
  és a tesztfuttató is lefuttatja őket
- a README pontosítva (a vezérlőcsatorna már nem unix socket, hanem hurokcím +
  jelszó), erőforrás-használat és tesztek fejezettel

+++ torrentdl/format.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import re
+import time
 
 DURATION_RE = re.compile(r"(?P<value>\d+(?:\.\d+)?)\s*(?P<unit>[smhd]?)", re.IGNORECASE)
 UNIT_SECONDS = {"s": 1, "m": 60, "h": 3600, "d": 86400, "": 60}
@@ -67,7 +68,7 @@ def human_time(seconds) -> str:
 
 def progress_bar(fraction: float, width: int = 30) -> str:
     fraction = max(0.0, min(1.0, float(fraction or 0)))
-    filled = int(round(fraction * width))
+    filled = round(fraction * width)
     return "[" + "#" * filled + "-" * (width - filled) + "]"
 
 
@@ -83,8 +84,6 @@ def eta_seconds(job: dict):
 
 def state_label(job: dict) -> str:
     """Az állapot rövid, magyar leírása (a szünet hátralévő idejével együtt)."""
-    import time
-
     state = job.get("state", "?")
     label = STATE_LABELS.get(state, state)
     if state == "seeding":
